pyproject.toml
[project]
name = "airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary==2.9.10",
    "apache-airflow==3.0.2",
    "apache-airflow[celery]",
    "asyncpg==0.30.0",
    "apache-airflow-providers-celery>=3.12.0",
    "apache-airflow[celery]",
    "flower==1.0.0",
    "redis==4.6.0",
    "mlflow==3.1.1"
]
